requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser>=6.0.11
feedgen==0.9.0
litellm==1.35.0
python-dotenv==1.0.0
lxml>=6.0.0
click==8.1.7
tqdm==4.66.1
selenium==4.15.2
webdriver-manager==4.0.1
qdrant-client==1.7.3
sentence-transformers==2.2.2
torch>=2.0.0
